# Perf: use fused Adam optimizer