@caic99 2024-12-09
Running with CPU is still possible with this parameter set.