@njzjz 2024-12-13
> For the profiling result shown below, this modification brings an 2.75x improvement on optimizer update (22ms vs. 8ms) and ~3% improvement for total speed up (922ms vs. 892ms). The benchmark case is training a DPA-2 Q3 release model. Please note that the absolute time may differs between steps.

It might be wrong to benchmark the time on the CPU since the GPU is asynchronous and is the bottleneck.