@caic99 2024-12-15
> which is the bottleneck

@njzjz What you've mentioned is the general case for NN itself. However, for optimizer update in deepmd-kit, the bottleneck is at the CPU side. You can refer to the "GPU SM efficiency" row from the profiling plot.